+++ modules/Grades/Grades.php
@@ -678,6 +678,10 @@ $LO_options['header'] = WrapTabs(
 
 echo '<br />';
 
+// @since 4.6 Navigate table inputs vertically using tab key.
+// @since 12.5 CSP remove unsafe-inline Javascript
+$LO_options['class'] = 'vertical-tab-navigation';
+
 if ( UserStudentID() )
 {
 	ListOutput(
@@ -702,25 +706,6 @@ else
 		$LO_options
 	);
 }
-
-// @since 4.6 Navigate form inputs vertically using tab key.
-// @link https://stackoverflow.com/questions/38575817/set-tabindex-in-vertical-order-of-columns
-?>
-<script>
-	function fixVerticalTabindex(selector) {
-		var tabindex = 1;
-		$(selector).each(function(i, tbl) {
-			$(tbl).find('tr').first().find('td').each(function(clmn, el) {
-				$(tbl).find('tr td:nth-child(' + (clmn + 1) + ') input').each(function(j, input) {
-					$(input).attr('tabindex', tabindex++);
-				});
-			});
-		});
-	}
-
-	fixVerticalTabindex('.list-wrapper .list tbody');
-</script>
-<?php
 
 echo $_REQUEST['assignment_id'] ? '<br /><div class="center">' . SubmitButton() . '</div>' : '';
 echo '</form>';